==> bloc3_machine_learning/walmart/src/1-WalMart_EDA.ipynb <==
# import des librairies

import pandas as pd

import plotly.graph_objects as go
import plotly.express as px
from plotly.subplots import make_subplots
import plotly.figure_factory as ff
import matplotlib.pyplot as plt

# Chargement des données

df = pd.read_csv("..\\data\\walmart_Store_Sales.csv")

## Exploration des données
### Affichage des premières lignes

df.head()

Pour bien comprendre les données, je vais chercher une date qui revient plusieurs fois. Cela me permettra de voir si les données sont liés au magasin (région) ou non

df['Date'].value_counts()
df_filtre = df[df['Date']=='19-10-2012']
df_filtre

Store - Numéro du magasin

Date - Semaine de vente (Toutes les dates sont des vendredis)

Weekly_Sales - Montant de vente de la semaine de vente pour le magasin (store)

Holiday_Flag - 1 si la semaine est une semaine de vacances, 0 sinon

Temperature - Temperature moyenne durant la semaine de vente (région du magasin)

Fuel_Price - Coût moyen du carburant durant la semaine de vente

CPI – Indice des prix à la consommation en vigueur (région du magasin)

Unemployment - Taux de chômage actuel (région du magasin)

# EDA et préprocessing

# Informations
print("Nombre de colonnes : {}".format(df.shape[1]))
print()

print("Nombre de lignes : {}".format(df.shape[0]))
print()

print("Affichage du dataset : ")
display(df.head())
print()

print("Statistiques basiques : ")
data_desc = df.describe(include="all")
display(data_desc)
print()

print("Pourcentage de valeurs manquantes : ")
Nb_Lignes = df.shape[0]
for column in df.columns:
    Nb_Null = df[column].isnull().sum()
    if Nb_Null != 0 :
        Pct_val_Null = round((Nb_Null/Nb_Lignes)*100,2)
        print(f"Dans la colonne {column}, il y a {Nb_Null} NULL soit {Pct_val_Null}%")

print("Type des donnees")
display (df.dtypes)

On peut noter que seule la colonne Store est intégralement renseigné. Le cible (Weekly_Sales) ne l'est pas toujours et je vais devoir supprimer les lignes.

df_cleaned = df.dropna(subset=['Weekly_Sales'])
df_cleaned.shape

On passe a 136 lignes après la suppression des 14 lignes sans valeur cible.

# je groupe par "store" pour determiner le nombre de na par "store"
df_na = df_cleaned.groupby("Store").apply(lambda group: group.isna().sum(), include_groups=False)

# Pour calculer le pourcentage de valeur manquante, je compte egalement le nombre de valeurs par "store"
df_value = df_cleaned.groupby("Store").apply(lambda group: group.notna().sum(), include_groups=False)

for store in df_na.index:
    na_counts = df_na.loc[store]
    value_counts = df_value.loc[store]
    total_na = na_counts.sum()
    total_value = value_counts.sum()
    ratio_na= round((total_na*100)/(total_value+total_na),2)
    print(f'le magasin {store} a {ratio_na}% de valeur manquante')

# Vu qu'aux USA, les vacances sont fixes par etat, on prendra la valeur la plus frequente ne sachant pas l'etat correspondant au store
mode_holiday = df_cleaned["Holiday_Flag"].mode()[0]
df_cleaned = df_cleaned.copy()
df_cleaned["Holiday_Flag"] = df_cleaned["Holiday_Flag"].fillna(mode_holiday)

Nb_Lignes = df_cleaned.shape[0]
for column in df_cleaned.columns:
    Nb_Null = df_cleaned[column].isnull().sum()
    if Nb_Null != 0 :
        Pct_val_Null = round((Nb_Null/Nb_Lignes)*100,2)
        print(f"Dans la colonne {column}, il y a {Nb_Null} NULL soit {Pct_val_Null}%")

Afin de rendre le champ Date plus facilement exploitable, je le convertis en 3 champs (Année, mois, Jour)

df_cleaned["Date"] = pd.to_datetime(df_cleaned["Date"], format="%d-%m-%Y")

df_cleaned["Year"] = df_cleaned["Date"].dt.year.astype("Int64")
df_cleaned["Month"] = df_cleaned["Date"].dt.month.astype("Int64")
df_cleaned["Day"] = df_cleaned["Date"].dt.day.astype("Int64")

df_cleaned.head()

Convertion du champ : 
 - Store en chaîne de caractère : variable qualitative
 - Holiday_Flag en chaine de caractère : variable qualitative

df_cleaned["Store"] = df_cleaned["Store"].astype(int).astype(str)
df_cleaned["Holiday_Flag"] = df_cleaned["Holiday_Flag"].astype(float).astype('int64').astype(str)



# Distribution des variables
## Contrôle des outliers

fig = go.Figure()

fig.add_trace(
    go.Box(
        x = df_cleaned['Temperature']))

fig.add_trace(
    go.Box(
        x = df_cleaned['Fuel_Price'],
        visible = False))

fig.add_trace(
    go.Box(
        x = df_cleaned['CPI'],
        visible = False))

fig.add_trace(
    go.Box(
        x = df_cleaned['Unemployment'],
        visible = False))


fig.update_layout(
        title = go.layout.Title(text = "Exploration differentes variables", x = 0.5),
        showlegend = False)

fig.update_layout(
    updatemenus = [go.layout.Updatemenu(
        active = 0,
        buttons = [
                    go.layout.updatemenu.Button(
                            label = "Temperature",
                            method = "update",
                            args = [{"visible" : [True, False, False, False]}]),
                    go.layout.updatemenu.Button(
                            label = "Prix du fuel",
                            method = "update",
                            args = [{"visible" : [False, True, False, False]}]),
                    go.layout.updatemenu.Button(
                            label = "Indice des Prix à la Consommation",
                            method = "update",
                            args = [{"visible" : [False, False, True, False]}]),
                    go.layout.updatemenu.Button(
                            label = "Chômage",
                            method = "update",
                            args = [{"visible" : [False, False, False, True]}])
                ]
    )]
)

A la vue des graphiques, nous allons suprimer les valeurs aberrantes du chômage uniquement

Moyenne_Unemployment = df_cleaned['Unemployment'].mean()
Ecart_type_Unemployment  = df_cleaned['Unemployment'].std()

mask_Chomage = ((df_cleaned['Unemployment']<Moyenne_Unemployment-3*Ecart_type_Unemployment)|(df_cleaned['Unemployment']>Moyenne_Unemployment+3*Ecart_type_Unemployment))
df_cleaned = df_cleaned[~mask_Chomage]

## Distribution des variables numériques

num_features = df_cleaned.select_dtypes(include=["int"]).columns

for feature in num_features:
    min_val = df_cleaned[feature].min()
    max_val = df_cleaned[feature].max()
    nb_bins = int(max_val) - int(min_val) +1
    bins = list(range(int(min_val), int(max_val) + 2))

    fig = px.histogram(df_cleaned, x=feature,nbins=nb_bins)
    fig.update_layout(title=f"Distribution de {feature}", title_x=0.5, xaxis_title=feature, yaxis_title="Nombre d'occurences")
    fig.update_xaxes(tickvals=bins)  # Force les ticks à être des entiers
    fig.show()


num_features = df_cleaned.select_dtypes(include=["float"]).columns

for feature in num_features:
    fig = px.histogram(df_cleaned, x=feature,nbins=30)
    fig.update_layout(title=f"Distribution de {feature}", title_x=0.5, xaxis_title=feature, yaxis_title="Nombre d'occurences")
    fig.show()


## Distribution des variables non numeriques

cat_features = df_cleaned.select_dtypes(include=["object"]).columns

for feature in cat_features:
    df_grouped = df_cleaned[feature].value_counts().reset_index()
    df_grouped.columns = [feature, "count"]

    fig = px.bar(df_grouped, x=feature, y="count", color_discrete_sequence=px.colors.qualitative.Pastel)
    fig.update_layout(title=f"Repartition de {feature}", title_x=0.5,
        xaxis_title=feature, xaxis=dict(tickformat="d"),
        yaxis_title="Nombre d'occurences")
    fig.show()

## Analyse temporelle

Répartition des données par année

fig = make_subplots(rows = 3, cols = 1)

years = df_cleaned["Year"].sort_values().dropna().unique().tolist()

for i in range(len(years)):
    mask = (df_cleaned["Year"] == years[i])
    df_year = df_cleaned[mask]

    fig.add_trace(go.Histogram(x = df_year["Month"], name = str(years[i]), nbinsx=12), row = i + 1, col = 1)

fig.update_layout(title = go.layout.Title(text = "Donnees mensuelles par annee", x = 0.5), height = 700)

fig.show()

Moyenne des ventes en fonction de la date sur les 3 ans

df_ventes = df_cleaned.groupby("Date")["Weekly_Sales"].mean().reset_index()
df_ventes["Mean_Sales"] = df_ventes["Weekly_Sales"].mean()
df_ventes.head()

df_ventes["Date"] = pd.to_datetime(df_ventes["Date"])

dates_ticks = pd.date_range(
    start=df_ventes["Date"].min(),
    end=df_ventes["Date"].max(),
    freq="3MS"  # Affichage des dates tous les trimestres
)

fig = go.Figure(
    data=go.Scatter(
        x=df_ventes["Date"],
        y=df_ventes["Weekly_Sales"],
        name="Ventes hebdomadaires"
    )
)

fig.add_trace(
    go.Scatter(
        x=df_ventes["Date"],
        y=df_ventes["Mean_Sales"],
        name="Moyenne des ventes"
    )
)

fig.update_layout(
    title=dict(text="Moyenne des ventes par jour", x=0.5),
    xaxis=dict(
        title="Date sur 3 ans",
        rangeslider=dict(visible=True),
        tickmode="array",  
        tickvals=dates_ticks,  
        tickformat="%b %Y",  
    ),
    showlegend=True 
)

fig.show()


On peut remarquer un impact de la période pour la moyenne des ventes. Cependannt, étant donné le nombre peu important des données, nous allons également vérifier l'impact des dates pour les magasins pour savoir si nous garderons ou non les données avec des dates manquantes.

## Analyse par magasin

Pour vérifier s'il est utile ou non de supprimer les lignes sans date sachant que nous souhaitons prédire les ventes par semaine, nous allons vérifier les moyennes de ventes avec et sans dates pour l'ensemble des magasins ains que de manière générale

# Moyenne par magasin (toutes lignes)
moyennes_par_magasin = df_cleaned.groupby('Store')['Weekly_Sales'].mean()

# Moyenne par magasin (dates nulles uniquement)
moyenne_nulles_par_magasin = df_cleaned[df_cleaned['Date'].isna()].groupby('Store')['Weekly_Sales'].mean()

# Moyenne generale
moyenne_generale = df_cleaned['Weekly_Sales'].mean()

# Moyenne generale pour dates nulles uniquement
moyenne_null_dates = df_cleaned[df_cleaned['Date'].isna()]['Weekly_Sales'].mean()

# Construction du DataFrame combine
df_comparatif = pd.DataFrame({
    'Toutes dates': moyennes_par_magasin,
    'Dates nulles uniquement': moyenne_nulles_par_magasin
})

# Tracer le barplot groupe
df_comparatif.plot(kind='bar', figsize=(10, 6), color=['lightgreen', 'salmon'], edgecolor='black')

# Ajouter lignes horizontales
plt.axhline(y=moyenne_generale, color='red', linestyle='--', label=f'Moyenne generale : {moyenne_generale:.2f}')
plt.axhline(y=moyenne_null_dates, color='orange', linestyle='--', label=f'Moyenne dates nulles : {moyenne_null_dates:.2f}')

# Titres et mise en forme
plt.title("Ventes moyennes par magasin")
plt.ylabel("Ventes moyennes")
plt.xlabel("Magasin")
plt.xticks(rotation=0)
plt.legend()
plt.tight_layout()
plt.show()


Je vais regarder si les magasins sont plus ou moins impactés par l'absence de date.

# Compter les dates non nulles et nulles par magasin
Store_dates = df_cleaned.groupby('Store')['Date'].agg(dates_non_nulles = lambda x: x.notna().sum(), dates_nulles = lambda x: x.isna().sum()).reset_index()

# Barplot
Store_dates.set_index('Store')[['dates_non_nulles', 'dates_nulles']].plot(
    kind='bar',
    stacked=False,
    figsize=(8, 5),
    color=['skyblue', 'salmon']
)

plt.title("Nombre de dates renseignees vs nulles par magasin")
plt.ylabel("Nombre d'occurrences")
plt.xlabel("Magasin")
plt.xticks(rotation=0)
plt.legend(title="Type de date")
plt.tight_layout()
plt.show()

L'impact de la suppression des lignes sans date, impactera particulièrement les magasins 10 et 11. Sur le premier graphique, on voit que la moyenne générale des ventes se rapproche de leur propre moyenne. Je décide donc de garder les lignes nulles afin de ne pas impacter trop l'absence des autres valeurs, particulièrement pour ces magasins.

## graphique bivarié et atrice de corrélation

fig = px.scatter_matrix(df_cleaned)

fig.update_layout(title = go.layout.Title(text = "Bivariate analysis", x = 0.5), showlegend = False, 
            autosize=False, height=1400, width = 1400)

fig.show()

Pas de corrélation frappante avec la variable cible.

corr_matrix = df_cleaned.corr().round(2)

fig = ff.create_annotated_heatmap(corr_matrix.values, x = corr_matrix.columns.tolist(), y = corr_matrix.index.tolist())
fig.show()

On observe qu'avec la variable cible `Weekly_Sales`, l'indice de correlation est intéressant pour :
 - `CPI`;
 - `Unemployment`;
 - `Temperature`;
 - `Store`.

# fin du nettoyage et sauvegarde du dataset nettoyé

Suppression des lignes si juste la date est non renseignée. On garde les autres valeurs qu'on remplacera par une valeur moyenne

# Masque pour les lignes où Date est Na
mask_date_na = df_cleaned['Date'].isna()

# Masque pour les lignes où l'un des aautres champ est na
mask_autre_na = df_cleaned[['Holiday_Flag','Temperature','CPI', 'Fuel_Price','Unemployment']].isna().any(axis=1)

# Combine les deux masques : on veut supprimer les lignes où les deux conditions sont vraies
mask_a_supp = mask_date_na & mask_autre_na
df_cleaned = df_cleaned[~mask_a_supp]


Suppression du champ Date que l'on a découpé et dont on ne servira plus

df_cleaned.drop(columns=['Date'],axis=1, inplace=True)

df_cleaned.to_csv("..\\data\\walmart_Store_Sales_Cleaned.csv")
