==> bloc3_machine_learning/walmart/src/3-Amelioration modele.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import  OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.linear_model import Ridge, Lasso,LinearRegression
from sklearn.metrics import r2_score,mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV


# chargement des données

df = pd.read_csv("..\\data\\Walmart_Store_sales_cleaned.csv")
df.head()

## stratify
Vu l'importance des "store", il est nécessaire de tester un stratify sur le store.

target = "Weekly_Sales"

X = df.drop(target, axis=1) 
Y = df[target]
display(X)
display(Y)

X_train, X_test, Y_train, Y_test = train_test_split(X, Y, test_size=0.2,random_state=20, stratify=X["Store"])
print(X_train.shape)
print(X_test.shape)

On voit une amélioration sur le test. Il existe toujours du sur-apprentissage mais bien moindre.

## Pré processing
- Pour les valeurs numériques
  - On impute les valeurs moyennes aux valeur absentes
  - On effectue un StandardScaler()  => Cela permet de mettre à l'échelle les données numériques en fonction de la moyenne et de l'écart type
- Pour les variables catégorielles
  - Pas d'imputation de valeur manquante car Store et Holiday_Flag sont tous renseignés
  - OneHotEncoder permet de convertir les catégories en vecteur binaire et on supprime la première occurence afin d'éviter la colinéarité 

numerical_columns = ["Temperature", "Fuel_Price", "CPI", "Unemployment", "Year", "Month", "Day"]
categorical_columns = ["Store", "Holiday_Flag"]

numeric_transformer = Pipeline(steps=[
    ("imputer", SimpleImputer(strategy="mean")),
    ("scaler", StandardScaler())
])

categorical_transformer = Pipeline(
    steps=[
    ("encoder", OneHotEncoder(drop="first", handle_unknown="ignore"))
    ])

feature_encoder = ColumnTransformer(
    transformers=[
        ("num", numeric_transformer, numerical_columns),
        ("cat", categorical_transformer, categorical_columns)
    ])

X_train = feature_encoder.fit_transform(X_train)
X_test = feature_encoder.transform(X_test) 

print("Train model...")
regressor = LinearRegression()
regressor.fit(X_train, Y_train)
print("...Done.")


print("R2 score training :", regressor.score(X_train, Y_train))
print("R2 score test :",  regressor.score(X_test, Y_test))


print("Predictions on training set...")
Y_train_pred = regressor.predict(X_train)
print("...Done.")

print("Predictions on test set...")
Y_test_pred = regressor.predict(X_test)
print("...Done.")

# Évaluation du modèle
mse = mean_squared_error(Y_test, Y_test_pred)
rmse = np.sqrt(mse)
mae = mean_absolute_error(Y_test, Y_test_pred)
r2_train = r2_score(Y_train, Y_train_pred)
r2_test = r2_score(Y_test, Y_test_pred)

print(f"RMSE: {rmse}")
print(f"MAE: {mae}")
print(f"R² train: {r2_train}")
print(f"R² test: {r2_test}")

scores_df = pd.read_csv("..\\data\\Walmart_Scores.csv")
new_rows = [{"model": "linear_regression_stratify", "R2_train": r2_train,"R2_test":r2_test,"RMSE":rmse,"MAE":mae}]
scores_df = pd.concat([scores_df, pd.DataFrame(new_rows)], ignore_index=True)

# Ridge

Ridge : Permet de réduire les coefficients des variables

Ridge sans alpha dans un remier temps (alpha = 1 par défaut)

ridge = Ridge()

# Entraîner le modèle
ridge.fit(X_train, Y_train)

print("Predictions sur le train...")
Y_train_pred = ridge.predict(X_train)
print("...Done.")

print("Predictions sur le test...")
Y_test_pred = ridge.predict(X_test)
print("...Done.")

print("R2 score sur le train : ", r2_score(Y_train, Y_train_pred))
print("R2 score sur le test : ", r2_score(Y_test, Y_test_pred))

# Évaluation du modèle
mse_ridge = mean_squared_error(Y_test, Y_test_pred)
rmse_ridge = np.sqrt(mse_ridge)
mae_ridge = mean_absolute_error(Y_test, Y_test_pred)
r2_train_ridge = r2_score(Y_train, Y_train_pred)
r2_test_ridge = r2_score(Y_test, Y_test_pred)

print(f"MSE: {mse_ridge}")
print(f"RMSE: {rmse_ridge}")
print(f"MAE: {mae_ridge}")
print(f"R²: {r2_test_ridge}")

L'écart entre le test et le train reste important et les résultats sont moins bons.

new_rows = [{"model": "ridge", "R2_train": r2_train_ridge,"R2_test":r2_test_ridge,"RMSE":rmse_ridge,"MAE":rmse_ridge}]
scores_df = pd.concat([scores_df, pd.DataFrame(new_rows)], ignore_index=True)

J'ajoute un hyper paramètre alpha

alphas = np.logspace(-4, 2, 50)
params = {'alpha': alphas}
gridsearch = GridSearchCV(ridge,params, cv=3, scoring="neg_mean_squared_error") # cv : the number of folds to be used for CV
gridsearch.fit(X_train, Y_train)
gridsearch.best_estimator_.get_params()

* La valeur alpha qui ressort est 0.087.


Y_train_pred = gridsearch.best_estimator_.predict(X_train)
Y_test_pred =  gridsearch.best_estimator_.predict(X_test)

print("R2 score on training set : ", r2_score(Y_train, Y_train_pred))
print("R2 score on test set : ", r2_score(Y_test, Y_test_pred))

mse_ridge2 = mean_squared_error(Y_test, Y_test_pred)
rmse_ridge2 = np.sqrt(mse_ridge2)
mae_ridge2 = mean_absolute_error(Y_test, Y_test_pred)
r2_train_ridge2 = r2_score(Y_train, Y_train_pred)
r2_test_ridge2 = r2_score(Y_test, Y_test_pred)

print(f"MSE: {mse_ridge2}")
print(f"RMSE: {rmse_ridge2}")
print(f"MAE: {mae_ridge2}")
print(f"R²: {r2_test_ridge2}")

new_rows = [{"model": "ridge_alpha", "R2_train": r2_train_ridge2,"R2_test":r2_test_ridge2,"RMSE":rmse_ridge2,"MAE":rmse_ridge2}]
scores_df = pd.concat([scores_df, pd.DataFrame(new_rows)], ignore_index=True)

La valeur en test grandment amélioré, cela réduit le sur-apprentissage même si l'écart reste presque de 0.04

# Lasso

Lasso : Lasso permet de pénaliser des variables jusqu'à les supprimer.

- Comme pour ridge, je commence sans valeur alpha.

lasso = Lasso()

# Entraîner le modèle
lasso.fit(X_train, Y_train)

print("Predictions on training set...")
Y_train_pred = lasso.predict(X_train)
print("...Done.")

print("Predictions on test set...")
Y_test_pred = lasso.predict(X_test)
print("...Done.")

print("R2 score on training set : ", r2_score(Y_train, Y_train_pred))
print("R2 score on test set : ", r2_score(Y_test, Y_test_pred))
# Évaluation du modèle
mse_lasso = mean_squared_error(Y_test, Y_test_pred)
rmse_lasso = np.sqrt(mse_lasso)
mae_lasso = mean_absolute_error(Y_test, Y_test_pred)
r2_train_lasso = r2_score(Y_train, Y_train_pred)
r2_test_lasso = r2_score(Y_test, Y_test_pred)

print(f"MSE: {mse_lasso}")
print(f"RMSE: {rmse_lasso}")
print(f"MAE: {mae_lasso}")
print(f"R²: {r2_test_lasso}")

Le test semble légèrement meilleur même le ridge avec hyper-paramètre

new_rows = [{"model": "lasso", "R2_train": r2_train_lasso,"R2_test":r2_test_lasso,"RMSE":rmse_lasso,"MAE":mae_lasso}]
scores_df = pd.concat([scores_df, pd.DataFrame(new_rows)], ignore_index=True)

- Je teste ensuite des valeurs de alpha.

alphas=np.logspace(-2, 4, 40)
params = {'alpha': alphas}

gridsearch = GridSearchCV(lasso, params, cv=3, scoring="neg_mean_squared_error")
gridsearch.fit(X_train, Y_train)
gridsearch.best_estimator_.get_params()


L'alpha 1701 ressort comme le meilleur. Il y a donc une forte pénalisation ce qui peut s'expliquer par le fait que la valeur cible est importante.

Y_train_pred = gridsearch.best_estimator_.predict(X_train)
Y_test_pred =  gridsearch.best_estimator_.predict(X_test)

print("R2 score on training set : ", r2_score(Y_train, Y_train_pred))
print("R2 score on test set : ", r2_score(Y_test, Y_test_pred))

mse_lasso2 = mean_squared_error(Y_test, Y_test_pred)
rmse_lasso2 = np.sqrt(mse_lasso2)
mae_lasso2 = mean_absolute_error(Y_test, Y_test_pred)
r2_train_lasso2 = r2_score(Y_train, Y_train_pred)
r2_test_lasso2 = r2_score(Y_test, Y_test_pred)

print(f"MSE: {mse_lasso2}")
print(f"RMSE: {rmse_lasso2}")
print(f"MAE: {mae_lasso2}")
print(f"R²: {r2_test_lasso2}")

new_rows = [{"model": "lasso_alpha"
"", "R2_train": r2_train_lasso2,"R2_test":r2_test_lasso2,"RMSE":rmse_lasso2,"MAE":mae_lasso2}]
scores_df = pd.concat([scores_df, pd.DataFrame(new_rows)], ignore_index=True)
scores_df.to_csv("../data/Walmart_Scores.csv", mode="w", index=False)
